optimisation spell system
déplacement du manacost dans les spells eux-mêmes et création d'une classe parent Spell (au lieu d'une interface).
Spells can now be rendered over the walls (AirboneSpells) or under (GroundedSpells) according to its render layer
ajout de john cena

+++ Assets/Arthur/Scripts/EnemyMovement.cs
@@ -35,7 +35,7 @@ public class EnemyMovement : MonoBehaviour
             }
             else
             {
-                int speed = enemyComponentReference.speed;
+                float speed = enemyComponentReference.speed;
                 transform.position = Vector3.MoveTowards(transform.position, path[targetIndex].position, speed * Time.deltaTime);
             }
         }
